First draft of demo for Pallas storylet system
Oh Jesus, where to start.

Okay so this was way more complicated than it really had to be. The goal here, in theory, was to make a very simple little demo to experiment with the storylet functionality of Twine. What resulted was a trip down the rabbit hole of Twine's documentation because I couldn't for the life of me figure out how to make the storylet macro run every time I changed the $demoGovernment variable. In the end I ended up looking for comparable situations in the forums. Found one in the end - it's inelegant, just a big nested series of if-else statements, but here we are. Core concept works!