Created intro and text effects
Got started on the first build of the game proper. I'm not 100% satisfied with the text-appearing system, but it will do for now. I'll get into this in a bit greater detail down the line, if there's time / need. In the meantime I'll ask around among my friends who are familiar with Twine about tips and tricks they might have.

My natural inclination is to write first and foremost and while I'm happy with this text so far, I'm going to need to do a lot of dry engineering and fill in the writing later because my peak effective hours for coding are a lot shorter than they are for writing.

Calling it for the night, but overjoyed to be back in the saddle and making a game again.