Added tutorial / project concept stages
Took the foot off the pedal for the writing for this one in order to focus on hammering out some underlying game mechanics. To wit, here's what I've come up with:

The game will proceed along a main quest which will be tracked with a mainPlotQuestStage variable. I'm drawing this from my fleeting attempts at modding Oblivion back in the day, which I understand uses a comparable quest log/flag system. It'll do as a spine.

The player's city at present has four principal resources: Strength (military strength, may rename to "might"), Piety (belief in the player-god), Stability (political continuity, the presence/absence of blood in the streets), and Prosperity (economic power, or lack thereof). Keeping these above a crisis threshold (2, at present) will keep the city ticking along, and the player's choices will have both narrative effects and increase/decrease these resources. This opens the interesting possibility of sacrificing a beloved character for the prosperity to keep the city out of the red, come to think of it, and the thought really warms my cold, black heart.

In between major plot beats, the player will "draw" events from a "deck" which is populated with storylets based on events in the game, be they boolean "this event has happened" flags or resources hitting certain critical thresholds. Currently the anticipated max is 5, though I may push higher. I'm also unsure if I want to have a "do well but not perfect" mechanic whereby having a city be too 'stable' results in a shift toward autocracy. We'll see. I'm eager to see what kinds of things I can do with this system, it seems amazingly modular and flexible and I'd love to get as many of these little storylets into the system as I can. Ah, the hubris at the start of the project.

Unrelated but important note: I'm a little annoyed at having to constantly reconfigure the exact delay timings for each passage, but at present I think building a separate config somewhere which is easily moddable is weirdly difficult in this particular version of Twine. Call it a "nice to have" for now unless it becomes truly burdensome. So if you notice that I'm redefining things a lot - I know, I know.

Also: the "Testbucket" is a technique I picked up working on Descent, I think it was. Essentially it's all of the basic config variables that I can copy and paste at the top of a scene to ensure that a scene can run independently. I could technically put some of this eg. formatting in a header, but for initializing variables, I need a lot of that stuff to-hand. It's inelegant, but I cut it all out before final release anyway, just on principle.

Switching over to some other tasks for now, but still feeling optimistic about this!