City Selector created
Got the city location selector to work. Main quest part 1 is now ⅓ complete. Feels good, even if so much of these growing pains have, in fact, felt bad!

Made a change to how I'm indicating rotating choices vs. time-advancing choices, namely that I'll include a cute little diamond next to all of the choices which advance time. This was brought about due to the fact that I cannot figure out how to get formatting changes on link-reveal text to not apply to all of the revealed text as well instead of just the link. I could either hammer my head against this for ages or just find a new solution and keep moving. Thanks Pathologic 2 for the inspiration!