Prologue Skeleton
Need to take a short break to attend to some personal matters but will be back at this later. The skeleton of the intro / drawing cards out of the act 1 deck seems to be functional. Next task will be to create templates for events so that I can make those a lot faster than I have been, then getting into basic mechanical testing this evening. It's coming together, improbably.

Additionally: added a unique economy for valley cities, since I am ditching the concept of a city-spirit, the valley city is a crossroads / holy site, a la Oxford, Santiago de Compostela, or Budapest.