Act 1 Skeleton Complete
Act 1 skeleton is finally (more or less) functional. I need to rewrite the introduction, which is sort of a shame because "Symmetry and self-symmetry, a spiral pulling up and up and up" is one of my favorite bits of writing in recent memory. Kill your darlings, I suppose.

I'm having quite a lot of fun with our Narrator. They're cranky, they're over these new kids with their cuneiform and their granaries and their emergent political formations. Truly "what if Homer was a cantankerous old rock star" - a theory incidentally which, despite my best efforts at publication, has yet to be accepted by any reputable classics journals.

I've done a quick playtest with my classmate, Florence, who obligingly clicked through the opening. She noted that it wasn't necessarily super obvious that the diamond indicates that the story should continue. I'll need to find a clearer form of visual differentiation.

Some process updates: I've created "Quickstart" profiles as a debug tool to make sure that the storylet decks work as designed. In essence, they're quick archetypes of a handful of choices collapsed into one place  - so far I have a communal hilltop mining city divided into districts and a decentralized valley pastoralist private-property-valuing community. I'll keep making these as I go through the acts in order to make the testing process more efficient.

There was also a significant difficulty in getting the storylets to work which bears explaining here. The (open-storylets:) macro pulls the first available storylet, which in this case is sorted by alphabetical order. Not ideal, necessarily, since I want the next storylet to be random.

My first solution was to use the (urgency:) macro. After a brief issue where I put the (urgency:) macro on the wrong line - nota bene, it has to go before the (storylet:) macro, not after - I had a deeply, profoundly frustrating bug where I kept getting an error that there are multiple (urgency:)s in the same passage. I could not for the life of me find it for the longest time, even after I removed all of the (urgency:) macros from - so I thought - all of the affected passages. In the end, I had indeed written an extra (urgency:) in a passage I have removed but not cut from the game altogether. I ended up finding an easier, more direct way to randomize the next storylet which will thankfully scale to the rest of the game.

Lots of progress! Pretty much all of it structural, but for the first time I can feel the shape of this thing looming in the distance. I've got this. I hope. Will do a bit more work on all this when I get home.